Nog enkele mogelijke 500 errors weggewerkt.

+++ src/main/java/be/ucll/campus/campus_app/service/CampusService.java
@@ -1,7 +1,10 @@
 package be.ucll.campus.campus_app.service;
 
+import be.ucll.campus.campus_app.exception.DuplicateResourceException;
+import be.ucll.campus.campus_app.exception.ResourceNotFoundException;
 import be.ucll.campus.campus_app.model.Campus;
 import be.ucll.campus.campus_app.repository.CampusRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,11 +14,18 @@ import java.util.Optional;
 public class CampusService {
     private final CampusRepository campusRepository;
 
+    @Autowired
     public CampusService(CampusRepository campusRepository) {
         this.campusRepository = campusRepository;
     }
 
     public Campus addCampus(Campus campus) {
+        Optional<Campus> bestaande = campusRepository.findById(campus.getNaam());
+
+        if (bestaande.isPresent()) {
+            throw new DuplicateResourceException("Campus met naam '" + campus.getNaam() + "' bestaat al.");
+        }
+
         return campusRepository.save(campus);
     }
 
@@ -33,6 +43,11 @@ public class CampusService {
 
 
     public void deleteCampus(String naam) {
+        Optional<Campus> campusOpt = campusRepository.findById(naam);
+        if (campusOpt.isEmpty()) {
+            throw new ResourceNotFoundException("Campus met naam '" + naam + "' niet gevonden.");
+        }
+
         campusRepository.deleteById(naam);
     }
